+++ app/src/dev/kotlin/xyz/lbres/trickcalculator/ui/devtools/DeveloperToolsDialog.kt
@@ -17,12 +17,12 @@ import androidx.lifecycle.ViewModelProvider
 import xyz.lbres.trickcalculator.BaseActivity
 import xyz.lbres.trickcalculator.R
 import xyz.lbres.trickcalculator.databinding.DialogDeveloperToolsBinding
-import xyz.lbres.trickcalculator.ext.view.gone
-import xyz.lbres.trickcalculator.ext.view.visible
 import xyz.lbres.trickcalculator.ui.BaseFragment
 import xyz.lbres.trickcalculator.ui.history.HistoryViewModel
 import xyz.lbres.trickcalculator.ui.settings.SettingsFragment
 import xyz.lbres.trickcalculator.ui.settings.SettingsViewModel
+import xyz.lbres.trickcalculator.utils.gone
+import xyz.lbres.trickcalculator.utils.visible
 
 /**
  * Dialog with various developer tools, only available in dev build flavor
