+++ app/src/espresso/kotlin/xyz/lbres/trickcalculator/utils/CallFunctionOnViewViewAction.kt
@@ -1,4 +1,4 @@
-package xyz.lbres.trickcalculator.ext.view
+package xyz.lbres.trickcalculator.utils
 
 import android.view.View
 import androidx.test.espresso.UiController
