--- app/src/main/kotlin/xyz/lbres/trickcalculator/ext/intrange/IntRangeExt.kt
@@ -1,8 +1,0 @@
-package xyz.lbres.trickcalculator.ext.intrange
-
-import xyz.lbres.kotlinutils.list.IntList
-import xyz.lbres.trickcalculator.SharedValues.random
-
-fun IntRange.appRandom(): Int = random(random)
-
-fun IntRange.appShuffled(): IntList = shuffled(random)
