--- app/src/main/kotlin/xyz/lbres/trickcalculator/ext/list/ListExt.kt
@@ -1,7 +1,0 @@
-package xyz.lbres.trickcalculator.ext.list
-
-import xyz.lbres.trickcalculator.SharedValues
-
-fun <E> List<E>.appRandom(): E = random(SharedValues.random)
-
-fun <E> List<E>.appShuffled(): List<E> = shuffled(SharedValues.random)
