+++ app/src/main/kotlin/xyz/lbres/trickcalculator/ui/history/generateRandomHistoryItem.kt
@@ -9,9 +9,8 @@ import xyz.lbres.kotlinutils.random.ext.nextBoolean
 import xyz.lbres.kotlinutils.random.ext.nextFromWeightedList
 import xyz.lbres.trickcalculator.SharedValues.random
 import xyz.lbres.trickcalculator.compute.isOperator
-import xyz.lbres.trickcalculator.ext.intrange.appRandom
-import xyz.lbres.trickcalculator.ext.list.appRandom
 import xyz.lbres.trickcalculator.utils.isNumber
+import xyz.lbres.trickcalculator.utils.seededRandom
 
 private typealias MStringList = MutableList<String>
 
@@ -60,7 +59,7 @@ fun generateRandomHistoryItem(): HistoryItem {
     // show error message instead of result
     val showError = random.nextBoolean(probabilityErrorMessage)
 
-    val length = (1..maxCompLength).appRandom()
+    val length = (1..maxCompLength).seededRandom()
     val computation = generateComputation(length)
 
     if (showError) {
@@ -90,7 +89,7 @@ private fun generateComputation(length: Int): StringList {
             )
             ef.toDecimalString(5)
         } else {
-            operators.appRandom()
+            operators.seededRandom()
         }
     }
 
@@ -122,7 +121,7 @@ private fun generateExactFraction(
     val negative = allowNegative && random.nextBoolean()
 
     val numeratorRange = random.nextFromWeightedList(weightedRanges)
-    var numerator = numeratorRange.random()
+    var numerator = numeratorRange.seededRandom()
     if (negative) {
         numerator *= -1
     }
@@ -131,7 +130,7 @@ private fun generateExactFraction(
     }
 
     val denominatorRange = random.nextFromWeightedList(weightedRanges)
-    val denominator = denominatorRange.random().ifZero { 1 }
+    val denominator = denominatorRange.seededRandom().ifZero { 1 }
     return ExactFraction(numerator, denominator)
 }
