+++ app/src/main/kotlin/xyz/lbres/trickcalculator/utils/appRandomUtils.kt
@@ -1,0 +1,16 @@
+package xyz.lbres.trickcalculator.utils
+
+import xyz.lbres.kotlinutils.list.IntList
+import xyz.lbres.trickcalculator.SharedValues.random
+
+/**
+ * IntRange methods using shared random as a seed
+ */
+fun IntRange.seededRandom(): Int = random(random)
+fun IntRange.seededShuffled(): IntList = shuffled(random)
+
+/**
+ * List methods using shared random as a seed
+ */
+fun <E> List<E>.seededRandom(): E = random(random)
+fun <E> List<E>.seededShuffled(): List<E> = shuffled(random)
